+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 2.3.0
+
+### Added
+
+- New `Nanook::Wallet#default_representative` method.
+- New `Nanook::Wallet#change_default_representative` method.
+- New `Nanook::Wallet#info` method.
+
 ## 2.2.0
 
 ### Added